Agrega template de configuración yaml y esqueleto de python script
