+++ .gitignore
@@ -1,2 +1,3 @@
 *.csv
 *.sh
+config.yml
